refactor: extract terminal launch
